Changes Made
For the changes made I have told the program to generate a random number and ask the player to guess what it is in a certain amount of tries. As the game progresses the program give the player feedback on their guess. If the player gets the answer right, it will congradulate them on it. If wrong it will tell them whether the guess was too low or too high until they guess the number correctly or they run out of tries in which the program will inform them that they have lost and tell them what the number was.